# docs/openapi.yaml
openapi: 3.0.0
info:
  title: LogBoard API
  description: API приложения для работы с логами "LogBoard".
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Authentication
    description: Эндпоинты для аутентификации.
  - name: Projects
    description: Эндпоинты для управления проектами.
  - name: Logs
    description: Эндпоинты для работы с логами.

security:
  - cookieAuth: []

paths:
  /register:
    post:
      summary: Регистрация пользователя
      description: Регистрирует нового пользователя с username и паролем.
      operationId: registerUser
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован.
        '400':
          description: Возвращает первую обнаруженную ошибку валидации данных.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь с таким username уже существует.

        '500':
          description: Внутренняя ошибка сервера.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /login:
    post:
      summary: Вход пользователя
      description: Аутентифицирует пользователя и выдает токены через HTTP-only cookies.
      operationId: loginUser
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
      responses:
        '200':
          description: Успешный вход. Токены устанавливаются через Set-Cookie заголовки.
          headers:
            Set-Cookie:
              schema:
                type: string
              example: access_token=eyJ...; HttpOnly; SameSite=Strict; Path=/; Max-Age=900
            X-Set-Cookie:
              schema:
                type: string
              description: Refresh token устанавливается через X-Set-Cookie заголовок (для избежания конфликтов с SameSite).
              example: refresh_token=eyJ...; HttpOnly; SameSite=Strict; Path=/; Max-Age=604800
        '400':
          description: Возвращает первую обнаруженную ошибку валидации данных.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверный логин или пароль.
        '500':
          description: Внутренняя ошибка сервера.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /refresh:
    post:
      summary: Обновление access token
      description: Использует refresh token из cookie для получения нового access token. Браузер автоматически отправляет cookie.
      operationId: refreshToken
      tags:
        - Authentication
      security: []
      responses:
        '200':
          description: Новый access token успешно получен и установлен через Set-Cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
              example: access_token=eyJ...; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=900
        '401':
          description: Refresh token истек или отсутствует.
        '500':
          description: Внутренняя ошибка сервера.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /logout:
    post:
      summary: Выход пользователя
      description: Очищает HTTP-only cookies с токенами.
      operationId: logoutUser
      tags:
        - Authentication
      security: []
      responses:
        '200':
          description: Пользователь успешно вышел. Cookies очищены.
          headers:
            Set-Cookie:
              schema:
                type: string
              example: access_token=; HttpOnly; SameSite=Strict; Path=/; Max-Age=0
            X-Set-Cookie:
              schema:
                type: string
              example: refresh_token=; HttpOnly; SameSite=Strict; Path=/; Max-Age=0
        '500':
          description: Внутренняя ошибка сервера.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /projects:
    post:
      summary: Создание нового проекта
      description: Создает новый проект с указанным названием. Пользователь, создающий проект, автоматически получает роль OWNER.
      operationId: createProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: Проект успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectCreateResponse'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Пользователь не авторизован
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Получение списка проектов пользователя
      description: Возвращает список всех проектов, в которых участвует текущий пользователь.
      operationId: getUserProjects
      tags:
        - Projects
      responses:
        '200':
          description: Список проектов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetProjectsResponse'
        '401':
          description: Пользователь не авторизован
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /projects/{projectId}:
    delete:
      summary: Удаление проекта
      description: Удаляет проект и все связанные с ним данные. Только владелец проекта может выполнить удаление.
      operationId: deleteProject
      tags:
        - Projects
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID проекта
      responses:
        '200':
          description: Проект успешно удален
        '401':
          description: Пользователь не авторизован
        '403':
          description: Недостаточно прав для выполнения операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Проект не найден
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /logs/search:
    post:
      summary: Поиск логов проекта
      description: Возвращает логи проекта с фильтрацией по уровню, тексту сообщения и временному периоду. Доступно всем участникам проекта.
      operationId: searchProjectLogs
      tags:
        - Logs
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogSearchRequest'
      responses:
        '200':
          description: Логи успешно получены.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogSearchResponse'
        '400':
          description: Ошибка валидации данных.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Пользователь не авторизован.
        '403':
          description: Пользователь не является участником проекта.
        '404':
          description: Проект не найден.
        '500':
          description: Внутренняя ошибка сервера.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /logs/timeline:
    post:
      summary: Таймлайн логов проекта
      description: Возвращает агрегированные счётчики логов по временной шкале на основе тех же фильтров, что и поиск. Обрабатывает всю выборку без пагинации. Доступно всем участникам проекта.
      operationId: getProjectLogsTimeline
      tags:
        - Logs
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogTimelineRequest'
      responses:
        '200':
          description: Таймлайн успешно получен.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimelineItem'
        '400':
          description: Ошибка валидации данных.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Пользователь не авторизован.
        '403':
          description: Пользователь не является участником проекта.
        '404':
          description: Проект не найден.
        '500':
          description: Внутренняя ошибка сервера.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: Access token передаётся через HTTP-only cookie "access_token".
    refreshCookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token передаётся через HTTP-only cookie "refresh_token". Используется только в эндпоинте /refresh.
  schemas:
    LogTimelineRequest:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
          description: ID проекта.
        level:
          type: array
          items:
            $ref: '#/components/schemas/LogLevel'
          description: Фильтр по уровням лога (можно указать несколько).
        message:
          type: string
          description: Полнотекстовый поиск по сообщению (регистронезависимо). Работает только если ClickHouse хранит поле message.
        from:
          type: string
          format: date-time
          description: Начало временного периода.
        to:
          type: string
          format: date-time
          description: Конец временного периода.
      required:
        - projectId
        - from
        - to
    LogSearchRequest:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
          description: ID проекта.
        level:
          type: array
          items:
            $ref: '#/components/schemas/LogLevel'
          description: Фильтр по уровням лога (можно указать несколько).
        message:
          type: string
          description: Полнотекстовый поиск по сообщению (регистронезависимо).
        from:
          type: string
          format: date-time
          description: Начало временного периода.
        to:
          type: string
          format: date-time
          description: Конец временного периода.
        size:
          type: integer
          default: 50
          description: Количество записей на странице.
        cursor:
          type: string
          format: date-time
          description: Курсор для получения следующей страницы (timestamp последнего лога из предыдущей выборки).
      required:
        - projectId
        - from
        - to
    LogSearchResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
        nextCursor:
          type: string
          format: date-time
          description: Курсор для получения следующей страницы (timestamp последнего лога из предыдущей выборки). Отсутствует если страниц больше нет.
        totalCount:
          type: integer
          description: Общее количество логов, соответствующих фильтру.
      required:
        - logs
        - totalCount
    LogLevel:
      type: string
      enum: [TRACE, DEBUG, INFO, WARN, ERROR]
    LogEntry:
      type: object
      properties:
        level:
          $ref: '#/components/schemas/LogLevel'
        message:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - level
        - message
        - timestamp
    TimelineItem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Начало интервала.
        totalCount:
          type: integer
          description: Общее количество логов за интервал.
        errorCount:
          type: integer
          description: Количество логов уровня ERROR за интервал.
        warnCount:
          type: integer
          description: Количество логов уровня WARN за интервал.
      required:
        - timestamp
        - totalCount
        - errorCount
        - warnCount
    GetProjectsResponse:
      type: array
      items:
        $ref: '#/components/schemas/GetProjectsResponseItem'
    GetProjectsResponseItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор проекта
        name:
          type: string
          description: Название проекта
        description:
          type: string
          description: Описание проекта
        createdAt:
          type: string
          format: date-time
          description: Дата и время создания проекта
        updatedAt:
          type: string
          format: date-time
          description: Дата и время последнего обновления проекта
        owner:
          type: string
          description: Username владельца проекта
      required:
        - id
        - name
        - createdAt
        - updatedAt
        - owner
    ProjectCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: Название проекта
          minLength: 1
          maxLength: 100
        description:
          type: string
          description: Описание проекта
          maxLength: 1000
      required:
        - name
    ProjectCreateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор проекта
    UserCredentials:
      type: object
      properties:
        username:
          type: string
          pattern: "^[a-zA-Z0-9_]{4,50}$"
          description: Уникальное имя пользователя, которое должно быть от 4 до 50 символов, только буквы, цифры и подчеркивания.
        password:
          type: string
          pattern: "^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{8,}$"
          description: Пароль пользователя, минимум 8 символов, должен включать хотя бы одну букву и одну цифру.
      required:
        - username
        - password
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Не удалось подключиться к базе данных